# Remove SampleControlPoint and DifficultyControlPoint from HitObject