--- a/osu.Game.Rulesets.Catch/Objects/BananaShower.cs
+++ b/osu.Game.Rulesets.Catch/Objects/BananaShower.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
+using System.Collections.Generic;
 using System.Threading;
+using osu.Game.Audio;
 using osu.Game.Rulesets.Judgements;
 using osu.Game.Rulesets.Objects.Types;
 
@@ -35,10 +37,11 @@ namespace osu.Game.Rulesets.Catch.Objects
             {
                 cancellationToken.ThrowIfCancellationRequested();
 
-                AddNested(new Banana(GetSampleInfo().Volume)
+                AddNested(new Banana
                 {
                     StartTime = time,
                     BananaIndex = i,
+                    Samples = new List<HitSampleInfo> { new Banana.BananaHitSampleInfo(GetSampleInfo().Volume) }
                 });
 
                 time += spacing;