--- a/osu.Game.Rulesets.Osu.Tests/Editor/TestSceneSliderStreamConversion.cs
+++ b/osu.Game.Rulesets.Osu.Tests/Editor/TestSceneSliderStreamConversion.cs
@@ -199,8 +199,7 @@ namespace osu.Game.Rulesets.Osu.Tests.Editor
                 Precision.AlmostEquals(circle.StartTime, time, 1)
                 && Precision.AlmostEquals(circle.Position, position, 0.01f)
                 && circle.NewCombo == startsNewCombo
-                && circle.Samples.SequenceEqual(slider.HeadCircle.Samples)
-                && circle.SampleControlPoint.IsRedundant(slider.SampleControlPoint);
+                && circle.Samples.SequenceEqual(slider.HeadCircle.Samples);
         }
 
         private bool sliderRestored(Slider slider)