… osu.Game.Rulesets.Mania/Resources/Testing/Beatmaps/convert-samples-expected-conversion.json
… osu.Game.Rulesets.Osu/Objects/Spinner.cs
… osu.Game.Rulesets.Osu/Objects/SpinnerBonusTick.cs
… osu.Game.Rulesets.Taiko/Beatmaps/TaikoBeatmapConverter.cs
… osu.Game.Rulesets.Taiko/Objects/DrumRoll.cs
… osu.Game.Rulesets.Taiko/Objects/TaikoStrongableHitObject.cs
… osu.Game.Rulesets.Taiko/UI/DrumSampleTriggerSource.cs
… osu.Game.Tests/Beatmaps/Formats/LegacyBeatmapDecoderTest.cs
… osu.Game.Tests/Editing/Checks/CheckMutedObjectsTest.cs
… osu.Game.Tests/Editing/TestSceneHitObjectComposerDistanceSnapping.cs
… osu.Game.Tests/Visual/Editing/TestSceneEditorClipboard.cs
… osu.Game.Tests/Visual/Editing/TestSceneEditorSaving.cs
… osu.Game.Tests/Visual/Editing/TestSceneHitObjectDifficultyPointAdjustments.cs
… osu.Game.Tests/Visual/Editing/TestSceneHitObjectSamplePointAdjustments.cs
… osu.Game.Tests/Visual/Gameplay/TestSceneGameplaySampleTriggerSource.cs
… osu.Game/Audio/HitSampleInfo.cs
… osu.Game/Beatmaps/ControlPoints/SampleControlPoint.cs
… osu.Game/Beatmaps/Formats/LegacyBeatmapDecoder.cs
… osu.Game/Beatmaps/Formats/LegacyBeatmapEncoder.cs
… osu.Game/Beatmaps/Legacy/LegacyContext.cs
… osu.Game/Context/ContextContainer.cs
… osu.Game/Context/IContext.cs
… osu.Game/Context/IHasContext.cs
… osu.Game/Rulesets/Edit/Checks/CheckMutedObjects.cs
… osu.Game/Rulesets/Edit/DistancedHitObjectComposer.cs
… osu.Game/Rulesets/Edit/PlacementBlueprint.cs
… osu.Game/Rulesets/Objects/Drawables/DrawableHitObject.cs
… osu.Game/Rulesets/Objects/HitObject.cs
… osu.Game/Rulesets/Objects/Legacy/ConvertHitObjectParser.cs
… osu.Game/Rulesets/Objects/Legacy/ConvertSlider.cs