The purpose is to clean up the `HitObject` class by not allowing any redundant information to be stored by the `HitObject`.
- The slider velocity and tick generating information has been moved to `Slider` and got a separate interface for access.
- Legacy information is now property of the `HitObject`.
- Sample control point information gets applied to the `Samples` directly,  making the sample control point information completely obsolete.

fixes #22171 #22132 #15618 #19483