@peppy 2023-05-02
Before dropping a review, huge thanks for taking on this effort. It's reassuring to have you both willing to understand and dive deep into the low-level structure of things surrounding hitobjects (which still have some work to get into a less-complex state).