@peppy 2023-05-03
I'm going to get these changes in. Pretty happy with how they are looking. I am sure we'll find a few weird issues post-merge but would rather keep moving forward.

For reference, here's benchmark results before and after:

`master`:

|                  Method |     Mean |     Error |    StdDev |     Gen0 |     Gen1 | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 6.744 ms | 0.1282 ms | 0.1199 ms | 343.7500 | 171.8750 |   3.78 MB |



|     Method | Count | WithBindableAccess |         Mean |       Error |      StdDev |     Gen0 |    Gen1 | Allocated |
|----------- |------ |------------------- |-------------:|------------:|------------:|---------:|--------:|----------:|
|  OsuCircle |     1 |              False |     178.9 ns |     1.47 ns |     1.30 ns |   0.0632 |  0.0002 |     728 B |
|   TaikoHit |     1 |              False |     316.8 ns |     2.97 ns |     2.63 ns |   0.0949 |  0.0005 |    1096 B |
| CatchFruit |     1 |              False |     126.7 ns |     1.46 ns |     1.29 ns |   0.0520 |       - |     600 B |
|  ManiaNote |     1 |              False |     114.8 ns |     1.62 ns |     1.51 ns |   0.0381 |       - |     440 B |
|  OsuCircle |     1 |               True |     308.5 ns |     6.16 ns |     5.46 ns |   0.1168 |  0.0010 |    1344 B |
|   TaikoHit |     1 |               True |     320.5 ns |     2.47 ns |     2.31 ns |   0.0949 |  0.0005 |    1096 B |
| CatchFruit |     1 |               True |     284.2 ns |     2.80 ns |     2.48 ns |   0.1192 |  0.0010 |    1376 B |
|  ManiaNote |     1 |               True |     130.1 ns |     2.58 ns |     2.87 ns |   0.0458 |       - |     528 B |

this PR:

|                  Method |     Mean |     Error |    StdDev |     Gen0 |     Gen1 | Allocated |
|------------------------ |---------:|----------:|----------:|---------:|---------:|----------:|
| BenchmarkBundledBeatmap | 5.766 ms | 0.0795 ms | 0.0744 ms | 328.1250 | 164.0625 |   3.66 MB |



|     Method | Count | WithBindableAccess |         Mean |       Error |      StdDev |     Gen0 |    Gen1 | Allocated |
|----------- |------ |------------------- |-------------:|------------:|------------:|---------:|--------:|----------:|
|  OsuCircle |     1 |              False |     174.9 ns |     1.88 ns |     1.76 ns |   0.0639 |  0.0002 |     736 B |
|   TaikoHit |     1 |              False |     307.9 ns |     2.44 ns |     2.28 ns |   0.0949 |  0.0005 |    1096 B |
| CatchFruit |     1 |              False |     125.7 ns |     1.57 ns |     1.47 ns |   0.0527 |  0.0002 |     608 B |
|  ManiaNote |     1 |              False |     109.5 ns |     0.43 ns |     0.40 ns |   0.0381 |  0.0001 |     440 B |
|  OsuCircle |     1 |               True |     288.6 ns |     2.65 ns |     2.48 ns |   0.1173 |  0.0010 |    1352 B |
|   TaikoHit |     1 |               True |     329.8 ns |     4.86 ns |     4.31 ns |   0.0949 |  0.0005 |    1096 B |
| CatchFruit |     1 |               True |     284.9 ns |     2.94 ns |     2.61 ns |   0.1202 |  0.0010 |    1384 B |
|  ManiaNote |     1 |               True |     125.5 ns |     2.15 ns |     2.01 ns |   0.0458 |       - |     528 B |