@OliBomby 2023-05-04
You can probably close these issues as of the merging of this PR: https://github.com/ppy/osu/issues/22132 https://github.com/ppy/osu/issues/15618 https://github.com/ppy/osu/issues/19483